Separate class in scripts directory
