+++ app/scripts/attachment.js
@@ -1,0 +1,6 @@
+export default class Attachment {
+  constructor(name, url) {
+    this.name = name;
+    this.url = url;
+  }
+}
