+++ app/scripts/background.js
@@ -1,6 +1,6 @@
 'use strict';
 
-import Board from './board';
+import MarkdownBuilder from './markdown-builder';
 
 chrome.browserAction.onClicked.addListener(tab => {
   let jsonUrl = /(https:\/\/trello.com\/b\/.*)\/.*/.exec(tab.url)[1] + '.json';
@@ -28,62 +28,3 @@ chrome.browserAction.onClicked.addListener(tab => {
   xmlHttp.open('GET', jsonUrl, true);
   xmlHttp.send();
 });
-
-class MarkdownBuilder {
-  static build(jsonData) {
-    let board = BoardBuilder.build(jsonData);
-
-    let markdownText = '## ' + board.name + '\n';
-    ['ToDo', 'Todo', 'Doing', 'Done'].forEach(listName => {
-      let lists = board.lists.find(list => list.name == listName)
-      if (lists == undefined) return;
-      markdownText += '\n### ' + listName + '\n\n';
-      lists.cards
-           .filter(card => !card.closed)
-           .forEach(card => markdownText += '- ' + card.name + '\n');
-    });
-
-    return markdownText;
-  }
-}
-
-class BoardBuilder {
-  static build(jsonData) {
-    let data = JSON.parse(jsonData);
-
-    let lists = data.lists.map(list => {
-      let filter_cards = data.cards.filter(card => card.idList == list.id);
-      let cards = filter_cards.map(card => {
-        let attachments = card.attachments.map(attachment => {
-          return new Attachment(attachment.name, attachment.url);
-        });
-        return new Card(card.name, attachments, card.closed);
-      });
-      return new List(list.name, cards);
-    });
-
-    return new Board(data.name, lists);
-  }
-}
-
-class Card {
-  constructor(name, attachments, closed) {
-    this.name = name;
-    this.attachments = attachments;
-    this.closed = closed;
-  }
-}
-
-class Attachment {
-  constructor(name, url) {
-    this.name = name;
-    this.url = url;
-  }
-}
-
-class List {
-  constructor(name, cards) {
-    this.name = name;
-    this.cards = cards;
-  }
-}
