+++ app/scripts/board.js
@@ -1,4 +1,4 @@
-class Board {
+export default class Board {
   constructor(name, lists) {
     this.name = name;
     this.lists = lists;
