+++ app/scripts/card.js
@@ -1,0 +1,7 @@
+export default class Card {
+  constructor(name, attachments, closed) {
+    this.name = name;
+    this.attachments = attachments;
+    this.closed = closed;
+  }
+}
