+++ app/scripts/list.js
@@ -1,0 +1,6 @@
+export default class List {
+  constructor(name, cards) {
+    this.name = name;
+    this.cards = cards;
+  }
+}
